+++ HISTORY.md
@@ -9,3 +9,14 @@
 
 # 0.0.4  / 2013-01-10
 * [Feature] functional::every, functional::some, functional::none
+
+# 0.0.5  / 2013-01-26
+* [Refactoring]
+
+# 0.0.6  / 2013-01-26
+* [Feature] functional::join
+  functional++/string.hpp
+  functional++/iterate.hpp
+  functional++/details.hpp
+  functional++/to_numeric.hpp
+  functional++/noncopyable.hpp